Made some temporary fixes

+++ Peace Treaty V1.2/Units/Provinces.cpp
@@ -72,7 +72,7 @@ int Provinces::findProvinceLevel()
 	for (int x: otherBuildingsLevels)
 		unitLevel += x;
 	
-	unitLevel /= resourceBuildingsLevels.size() + otherBuildingsLevels.size();
+	unitLevel /= (int) resourceBuildingsLevels.size() + (int)otherBuildingsLevels.size();
 	return unitLevel;
 }
